add get last token id
