+++ packages/contracts/src/WebmaToken.sol
@@ -2,12 +2,23 @@
 pragma solidity ^0.8.14;
 
 import "../lib/openzeppelin-contracts/contracts/token/ERC721/ERC721.sol";
-
+import "../lib/openzeppelin-contracts/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
+import "../lib/openzeppelin-contracts/contracts/utils/Counters.sol";
 
 contract WebmaToken is ERC721 {
+    using Counters for Counters.Counter;
+
+    Counters.Counter private _tokenIdCounter;
+
     constructor() ERC721("WebmaToken", "WMT") {}
 
-    function mint(address to, uint256 tokenId) public {
-        _mint(to, tokenId);
+    function mint(address to) public {
+        uint256 tokenId = _tokenIdCounter.current();
+        _tokenIdCounter.increment();
+        _safeMint(to, tokenId);
+    }
+
+    function getLastTokenId() external view returns (uint256) {
+        return _tokenIdCounter.current() - 1;
     }
 }
