+++ packages/frontend/.env.example
@@ -1,3 +1,3 @@
-NEXT_PUBLIC_WEBMA_SWAP_ADDRESS=0x373be0f474ed6717d9e603f41005531b7a7c1747
-NEXT_PUBLIC_WEBMA_TOKEN_ADDRESS=0xdfaa42ceacc12c908f968fb13b57d7e8265ff27f
+NEXT_PUBLIC_WEBMA_SWAP_ADDRESS=0x2e0f3680637e5fb165236deb547c59684701b2d8
+NEXT_PUBLIC_WEBMA_TOKEN_ADDRESS=0x41165a329487c4b530c6002c8be9d538924e54d0
 NEXT_PUBLIC_CHAIN_ID=5
