+++ packages/frontend/hooks/useContract.ts
@@ -1,7 +1,7 @@
-import { useContractWrite, usePrepareContractWrite } from "wagmi";
-import { abi as erc20Abi } from "./abi/ERC20";
-import { abi as WebmaTokenAbi } from "./abi/WebmaToken";
-import { abi as webmaSwapAbi } from "./abi/WebmaSwap";
+import { useContractWrite, usePrepareContractWrite, useContractReads } from "wagmi";
+import { abi as erc20Abi } from "../abi/ERC20";
+import { abi as WebmaTokenAbi } from "../abi/WebmaToken";
+import { abi as webmaSwapAbi } from "../abi/WebmaSwap";
 
 export const useApproveERC20 = (address: any, amount: any) => {
   const { config } = usePrepareContractWrite({
@@ -56,4 +56,16 @@ export const useFulfill = (tokenId: any) => {
     args: [tokenId],
   });
   return useContractWrite(config);
+};
+
+export const useGetLastTokenId = () => {
+  return useContractReads({
+    contracts: [
+      {
+        address: process.env.NEXT_PUBLIC_WEBMA_TOKEN_ADDRESS,
+        abi: WebmaTokenAbi,
+        functionName: "getLastTokenId",
+      },
+    ],
+  });
 };
